Major changes in room_controller.
1. Use single connection and single channel created when server start.
According some documents on web. A channel is a TCP connection. So one or a few
channels for each tornado server is reasonable...

2. No more exchange declaring. Since the room should exist before any user try to enter.
So declaring exchange every time doesn't make sense.

3. Specified consumer_tag for basic_consume instead of using pika's auto generated name...
I believe there is a trivial bug in pika's channel. Now the tag is thread lock protected..
But seems useless because we will use multi-processes only

4. add "declare_queue_only" parameter for Channel(object), because when user entering a 
room, we created a durable queue for listening boardcase msg. But we don't want to consume
the message at this moment 

5. Most of the RequestHandler won't put itself to Channel(object) because the channel will
try to send session['messages'] in case of that. Which is only used for BoardListenMessageHandler

6. Channel(object) will clear all the event handler when channel.close() is called. 
So that normal RequestHandler should call self.finish() directly, except for BoardListenMessageHandler.
Also added a channel.closing variable to ensure a consumer_tag will be cancel only once

7. Passed queue_name to Channel(object) is no longer used for non-durable queue.
Because:
	a.we won't use the queue again
	b.For example, user called "sit" in very short time,
	  if we have a shared queue. The reply message for second request will be receive by
	  the first RequestHandler. Which will make second RequestHandler last forever 
	c. for durable queue, we really want a shared queue in this case.


8. Except enter, other RequestHandler is not tested. Bugs may exist



git-svn-id: http://10.0.0.3/svn/poker@120 f0252c50-933b-4dd8-9fc7-05cab60132cc